Fixup merge

+++ persistence/bitmagnet.go
@@ -129,7 +129,7 @@ func (b *bitmagnet) GetNumberOfQueryTorrents(query string, epoch int64) (uint64,
 	return 0, nil
 }
 
-func (b *bitmagnet) QueryTorrents(query string, epoch int64, orderBy OrderingCriteria, ascending bool, limit uint64, lastOrderedValue *float64, lastID *uint64) ([]TorrentMetadata, error) {
+func (b *bitmagnet) QueryTorrents(query string, epoch int64, orderBy OrderingCriteria, ascending bool, limit uint64, lastOrderedValue *uint64, lastID *uint64) ([]TorrentMetadata, error) {
 	return nil, errors.New("query not supported")
 }
